Rename hash-utils.* to hash_utils.*
