+++ serum.files
@@ -24,12 +24,12 @@ libserum/core/setup/platform-setup.h
 libserum/core/setup/special-treatment-for-windows.h
 libserum/core/setup/types.h
 libserum/core/signatures.h
-libserum/crypto/hash-utils.c
-libserum/crypto/hash-utils.h
 libserum/crypto/hash.c
 libserum/crypto/hash.h
 libserum/crypto/hash_registry.c
 libserum/crypto/hash_registry.h
+libserum/crypto/hash_utils.c
+libserum/crypto/hash_utils.h
 libserum/crypto/hashing/md5.c
 libserum/crypto/hashing/md5.h
 libserum/crypto/hashing/ripemd160.c
